drools integration
