+++ src/main/java/com/livelabdrools/application/Application.java
@@ -1,12 +1,10 @@
 package com.livelabdrools.application;
 
+import com.livelabdrools.configuartion.DroolsConfiguration;
 import com.livelabdrools.configuartion.SpringELConfiguration;
 import com.livelabdrools.mapper.DataMapper;
-import com.livelabdrools.mapper.PersonMapper;
 import com.livelabdrools.model.Person;
 import com.livelabdrools.rule.RuleEngine;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.beans.factory.annotation.Qualifier;
 import org.springframework.boot.autoconfigure.EnableAutoConfiguration;
 import org.springframework.context.ApplicationContext;
 import org.springframework.context.annotation.AnnotationConfigApplicationContext;
@@ -39,14 +37,16 @@ public class Application {
 
     public static void main(String[] args) {
         ApplicationContext context = new AnnotationConfigApplicationContext(SpringELConfiguration.class);
+//        ApplicationContext context = new AnnotationConfigApplicationContext(DroolsConfiguration.class);
         Application app = (Application)context.getBean("app");
-        app.processData("C:\\Users\\674448\\Desktop\\personinput.psv"/*args[0]*/);
+        app.processData("C:\\Users\\690257\\Desktop\\personinput.psv"/*args[0]*/);
     }
 
     public void processData(String inputFile) {
         File file=new File(inputFile);
         List<Person> objToProcess = dataMapper.getData(file);
-        /*List<Person> processedObj = */ruleEngine.processData(objToProcess);
+        List<Person> processedObj =        ruleEngine.processData(objToProcess);
+        System.out.println(processedObj);
         //processedObj
     }
