+++ src/main/java/com/livelabdrools/configuartion/DroolsConfiguration.java
@@ -5,8 +5,8 @@ import com.livelabdrools.mapper.DataMapper;
 import com.livelabdrools.mapper.PersonMapper;
 import com.livelabdrools.mapper.RuleMapper;
 import com.livelabdrools.model.Rule;
+import com.livelabdrools.rule.DroolsRuleEngine;
 import com.livelabdrools.rule.RuleEngine;
-import com.livelabdrools.rule.SpringRuleEngine;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
 
@@ -14,21 +14,16 @@ import java.io.File;
 import java.util.List;
 
 @Configuration
-public class SpringELConfiguration {
-
-    private DataMapper getPersonMapper() {
-        return new PersonMapper();
-    }
+public class DroolsConfiguration { private DataMapper getPersonMapper() {
+    return new PersonMapper();
+}
 
     private RuleEngine getRuleEngine() {
-        DataMapper dataMapper = new RuleMapper();
-        ClassLoader classLoader = SpringELConfiguration.class.getClassLoader();
-        List<Rule> ruleList = dataMapper.getData(new File(classLoader.getResource("SpringEL_Rule.xlsx").getFile()));
-        RuleEngine ruleEngine = new SpringRuleEngine(ruleList);
+        RuleEngine ruleEngine = new DroolsRuleEngine();
         return ruleEngine;
     }
 
-    @Bean("app")
+    @Bean("appDrools")
     public Application getApplication() {
         Application app = new Application();
         app.setDataMapper(getPersonMapper());
