--- src/main/java/com/livelabdrools/rule/SpringELRuleEngine.java
@@ -1,15 +1,0 @@
-package com.livelabdrools.rule;
-
-import com.livelabdrools.model.Person;
-import com.livelabdrools.model.RuleFact;
-
-import java.util.List;
-
-public interface SpringELRuleEngine {
-
-    public void processData (List<Person> personList);
-    public String parseInputRule(List<RuleFact> ruleFactList);
-
-
-
-}
