+++ src/main/resources/META-INF/kmodule.xml
@@ -1,0 +1,6 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<kmodule xmlns="http://jboss.org/kie/6.0.0/kmodule">
+   <kbase name="rules" >
+        <ksession name="ksession-rule" default="true"/>
+    </kbase>
+</kmodule>
