+++ src/main/resources/META-INF/maven/pom.properties
@@ -1,0 +1,3 @@
+<groupId>com.livelab.drools.poc</groupId>
+<artifactId>livelab_drools_poc</artifactId>
+<version>1.0-SNAPSHOT</version>
